Refactor: Renamed folder, updated field name, and cleaned up unused imports
