+++ rust/main/config/mainnet_config.json
@@ -4055,6 +4055,22 @@
       "index": {
         "from": 26847587
       }
+    },
+    "ton": {
+      "name": "ton",
+      "tonApiCenter": [
+        {
+          "http": "https://toncenter.com/api/v3/",
+          "apiToken": "YOUR_API_TOKEN_HERE"
+        }
+      ],
+      "nativeToken": {
+        "name": "Toncoin",
+        "symbol": "TON",
+        "decimals": 9
+      },
+      "chainId": 0,
+      "isTestnet": false
     }
   },
   "defaultRpcConsensusType": "fallback"
